@@ -1,13 +1,13 @@
-import React, { Fragment } from 'react'
+import React from 'react'
 import Search from '../users/Search';
 import Users from '../users/Users';
 
 const Home = () => {
     return (
-        <Fragment>
+      <>
         <Search />
         <Users />
-      </Fragment>
+      </>
     )
 }
 
