@@ -1,19 +1,19 @@
-import React, { useEffect, Fragment, useContext } from 'react'
+import React, { useEffect, useContext } from 'react'
 import Spinner from '../layout/Spinner'
 import Repos from '../repos/Repos'
-import { Link } from 'react-router-dom'
+import { Link, useParams } from 'react-router-dom'
 import GithubContext from '../../context/github/githubContext'
 
 
-const User = ({ match }) => {
+const User = () => {
+    const { id } = useParams();
+    const githubContext = useContext(GithubContext);
 
-    const githubContext = useContext(GithubContext)
-
-    const { loading, getUser, user, getUserRepos, repos } = githubContext
+    const { loading, getUser, user, getUserRepos, repos } = githubContext;
 
     useEffect(() => {
-        getUser(match.params.login);
-        getUserRepos(match.params.login);
+        getUser(id);
+        getUserRepos(id);
         // eslint-disable-next-line
     }, []);
 
@@ -31,48 +31,47 @@ const User = ({ match }) => {
         public_gists,
         hireable
     } = user;
-
-    if(loading) return <Spinner />;
     
     return (
-        <Fragment>
-            <Link to='/' className="btn btn-light">Back to Search</Link>
-            Hireable: {''}
-            {hireable ? (<i className="fas fa-check text-success"></i>) :  (<i className="fas fa-times text-danger"></i>)}
-            <div className="card grid-2">
-                <div className="all-center">
-                    <img src={avatar_url} className='round-img' alt="" style={{ width: '150px' }}/>
-                    <h1>{name}</h1>
-                    {location && <Fragment> Location: {location} </Fragment>}
+        loading ? <Spinner /> : (
+            <>
+                <Link to='/' className="btn btn-light">Back to Search</Link>
+                Hireable: <a href={html_url} target="_blank" rel="noopener noreferrer"><i className={`fas fa-${hireable ? 'check' : 'times'} text-${hireable ? 'success' : 'danger'}`}></i></a>
+                <div className="card grid-2">
+                    <div className="all-center">
+                        <img src={avatar_url} className='round-img' alt="" style={{ width: '150px' }}/>
+                        <h1>{name}</h1>
+                        {location && <> Location: {location} </>}
+                    </div>
+                    <div>
+                        {bio && <>
+                            <h3>Bio</h3>
+                        <p>{bio}</p>
+                        </>}
+                        <a href={html_url} target="_blank" rel="noopener noreferrer" className='btn btn-dark my-1'>Visit GitHub Profile</a>
+                        <ul>
+                            <li>{login && <> Username: {login}</>}</li>
+                            <li>{blog && <> Website: {blog}</>}</li>
+                        </ul>
+                    </div>
                 </div>
-                <div>
-                    {bio && <Fragment>
-                        <h3>Bio</h3>
-                    <p>{bio}</p>
-                    </Fragment>}
-                    <a href={html_url} target="_blank" rel="noopener noreferrer" className='btn btn-dark my-1'>Visit GitHub Profile</a>
-                    <ul>
-                        <li>{login && <Fragment> Username: {login}</Fragment>}</li>
-                        <li>{blog && <Fragment> Website: {blog}</Fragment>}</li>
-                    </ul>
+                <div className="card text-center">
+                    <div className="badge badge-success">
+                        Followers: {followers}
+                    </div>
+                    <div className="badge badge-light">
+                        Following: {following}
+                    </div>
+                    <div className="badge badge-danger">
+                        Public Repos: {public_repos}
+                    </div>
+                    <div className="badge badge-dark">
+                        Public Gists: {public_gists}
+                    </div>
                 </div>
-            </div>
-            <div className="card text-center">
-                <div className="badge badge-success">
-                    Followers: {followers}
-                </div>
-                <div className="badge badge-light">
-                    Following: {following}
-                </div>
-                <div className="badge badge-danger">
-                    Public Repos: {public_repos}
-                </div>
-                <div className="badge badge-dark">
-                    Public Gists: {public_gists}
-                </div>
-            </div>
-        <Repos repos={repos}/>
-        </Fragment>
+                <Repos repos={repos}/>
+            </>
+        )
     )
 }
 
